feat(editProfile): refactor code structure to adhere to best practice
